Title: 특정 문자로 시작하는 문자열, Time: 141ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240617/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -18,7 +18,7 @@ public class Main {
                 sum += arr[i].length();
             }
         }
-        avg = sum / count;
+        avg = (double)sum / count;
         System.out.printf("%d %.2f", count, avg);
     }
 }
